[as6812-32x] Add system eeprom support

+++ components/all/platform-config/x86-64-accton-as6812-32x-r0/src/boot/x86-64-accton-as6812-32x-r0-devices.sh
@@ -8,6 +8,16 @@
 # </bsn.cl>
 ############################################################
 
+#
+# The IDPROM device is initialized last as part of this script.
+#
+# Assume that if the IDPROM device already exists then we
+# have already executed properly.
+#
+if [ -f /sys/devices/pci0000:00/0000:00:13.0/i2c-1/1-0057/eeprom ]; then
+    exit 0
+fi
+
 ########### initialize I2C bus 0 ###########
 # initialize CPLD
 echo as6812_32x_cpld1 0x60 > /sys/bus/i2c/devices/i2c-0/new_device
@@ -65,5 +75,8 @@ echo lm75 0x48 > /sys/bus/i2c/devices/i2c-38/new_device
 echo lm75 0x49 > /sys/bus/i2c/devices/i2c-39/new_device
 echo lm75 0x4a > /sys/bus/i2c/devices/i2c-40/new_device
 echo lm75 0x4b > /sys/bus/i2c/devices/i2c-41/new_device
+
+# IDPROM
+echo 24c02 0x57 > /sys/devices/pci0000:00/0000:00:13.0/i2c-1/new_device
 
 exit 0
